backend/pyproject.toml
[project]
name = "juggertube"
version = "0.0.0"
dependencies = [
    "alembic",
    "autoflake",
    "autopep8",
    "celery",
    "cryptography",
    "flask",
    "flask_caching",
    "flask_caching",
    "flask_compress",
    "flask_cors",
    "flask_inputfilter",
    "flask_jwt_extended",
    "flask_limiter",
    "flask_migrate",
    "flask_sqlalchemy",
    "flask_talisman",
    "google-api-python-client",
    "gunicorn",
    "isort",
    "openpyxl",
    "pandas",
    "pillow",
    "pymysql",
    "pyOpenSSL",
    "pytest",
    "pytest_cov",
    "pytest_env",
    "pyyaml",
    "python-telegram-bot>=20.7",
    "redis",
    "requests",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["config", "DataDomain", "ExternalApi", "Infrastructure"]

[tool.isort]
profile = 'black'
skip = ['venv', '.venv', '*.md']
line_length = 90

[tool.flake8]
exclude = ['__init__.py', 'venv', '.venv', '*.md']
max-line-length = 90
